+++ Library/src/entities/SolidEntity.cpp
@@ -213,6 +213,7 @@ std::vector<Renderable> SolidEntity::Render()
     {
         Renderable item;
         item.type = RenderableType::SOLID;
+        item.classId = getSegmentationClassId();
         item.materialName = mat.name;
         
         if(dm == DisplayMode::GRAPHICAL && graObjectId >= 0)